requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
zhipuai>=2.1.2
zai-sdk>=0.2.0
llama-index>=0.14.13
llama-index-vector-stores-chroma>=0.1.10
llama-index-embeddings-openai>=0.1.10
llama-index-embeddings-zhipuai>=0.1.10
chromadb>=0.4.22
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiofiles>=23.2.1
sqlalchemy>=2.0.25
alembic>=1.13.1
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic[email]>=2.6.0
python-docx>=1.1.0
markdown>=3.7.0
